requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
Flask-Cors==4.0.0
